requirements.txt
amqp==5.2.0
argcomplete==2.0.0
asgiref==3.7.2
Beaker==1.12.1
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
blessings==1.7
blivet==3.8.2
blivet-gui==2.5.0
Brlapi==0.8.5
celery==5.3.6
certifi==2024.2.2
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
chess==1.10.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
conda==23.9.0
conda-package-handling==2.1.0
conda_package_streaming==0.7.0
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.8.0
Django==4.2.7
dnf==4.19.0
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
frozendict==2.3.8
googletrans==3.0.0
gpg==1.20.0
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
humanize==3.13.1
hyperframe==5.2.0
idna==2.10
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.3
kombu==5.3.5
langtable==0.0.65
libcomps==0.1.20
libdnf==0.73.0
lxml==4.9.3
Mako==1.2.3
MarkupSafe==2.1.3
mwparserfromhell==0.6.6
nftables==0.1
numpy==1.24.4
olefile==0.46
packaging==23.1
Paste==3.5.3
pbkdf2==1.3
pexpect==4.8.0
pid==2.2.3
pillow==10.2.0
pluggy==1.2.0
ply==3.11
productmd==1.38
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pwquality==1.4.5
py-cpuinfo==9.0.0
pycairo==1.25.1
pycosat==0.6.4
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
pygame==2.5.2
Pygments==2.15.1
PyGObject==3.46.0
pykickstart==3.48
pyOpenSSL==23.2.0
pyparted==3.13.0
pyserial==3.5
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
pyudev==0.24.1
pywikibot==9.0.0
pyxdg==0.27
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rfc3986==1.5.0
rpm==4.19.1.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scapy==2.5.0
scikit-learn==1.4.1.post1
scipy==1.12.0
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.6.0
soupsieve==2.5
SpeechRecognition==3.10.1
sqlparse==0.4.4
systemd-python==235
Tempita==0.5.2
threadpoolctl==3.4.0
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
wifi==0.3.8
wikipedia==1.4.0
zstandard==0.21.0
